requirements.txt
google-cloud-bigquery==3.*
